Rename associated-token-address -> associated-token-account-client
