+++ Cargo.toml
@@ -3,9 +3,9 @@ split-debuginfo = "unpacked"
 
 [workspace]
 members = [
+  "associated-token-account/client",
   "associated-token-account/program",
   "associated-token-account/program-test",
-  "associated-token-address",
   "binary-option/program",
   "binary-oracle-pair/program",
   "examples/rust/cross-program-invocation",
