+++ associated-token-account/client/Cargo.toml
@@ -1,15 +1,15 @@
 [package]
-name = "spl-associated-token-address"
+name = "spl-associated-token-account-client"
 version = "1.0.0"
-description = "Solana Program Library Associated Token Address"
+description = "Solana Program Library Associated Token Account Client"
 authors = ["Solana Labs Maintainers <maintainers@solanalabs.com>"]
 repository = "https://github.com/solana-labs/solana-program-library"
 license = "Apache-2.0"
 edition = "2021"
 
 [dependencies]
-solana-inline-spl = "2.1.0"
-solana-program = "2.1.0"
+solana-inline-spl = "2.0.3"
+solana-program = "2.0.3"
 
 [package.metadata.docs.rs]
 targets = ["x86_64-unknown-linux-gnu"]
