+++ associated-token-account/client/src/lib.rs
@@ -1,0 +1,8 @@
+//! Client crate for interacting with the spl-associated-token-account program
+#![deny(missing_docs)]
+#![forbid(unsafe_code)]
+
+pub mod address;
+pub mod instruction;
+
+solana_program::declare_id!("ATokenGPvbdGVxr1b2hvZbsiqW5xWH25efTNsLJA8knL");
