+++ associated-token-account/program/Cargo.toml
@@ -17,7 +17,7 @@ borsh = "1.5.1"
 num-derive = "0.4"
 num-traits = "0.2"
 solana-program = "2.0.3"
-spl-associated-token-address = { version = "1.0.0", path = "../../associated-token-address" }
+spl-associated-token-account-client = { version = "1.0.0", path = "../client" }
 spl-token = { version = "6.0", path = "../../token/program", features = [
   "no-entrypoint",
 ] }
