+++ associated-token-account/program/src/instruction.rs
@@ -1,17 +1,10 @@
 //! Program instructions
 
-use {
-    crate::id,
-    assert_matches::assert_matches,
-    borsh::{BorshDeserialize, BorshSchema, BorshSerialize},
-    solana_program::{
-        instruction::{AccountMeta, Instruction},
-        pubkey::Pubkey,
-    },
-    spl_associated_token_address::get_associated_token_address_with_program_id,
-};
+use borsh::{BorshDeserialize, BorshSchema, BorshSerialize};
+pub use spl_associated_token_account_client::intruction::*;
 
 /// Instructions supported by the AssociatedTokenAccount program
+#[deprecated(since = "4.1.0", note = "Use instructions in `spl-associated-token-account-client` crate instead.")]
 #[derive(Clone, Debug, PartialEq, BorshDeserialize, BorshSerialize, BorshSchema)]
 pub enum AssociatedTokenAccountInstruction {
     /// Creates an associated token account for the given wallet address and
@@ -55,108 +48,4 @@ pub enum AssociatedTokenAccountInstruction {
     ///      account
     ///   6. `[]` SPL Token program
     RecoverNested,
-}
-
-fn build_associated_token_account_instruction(
-    funding_address: &Pubkey,
-    wallet_address: &Pubkey,
-    token_mint_address: &Pubkey,
-    token_program_id: &Pubkey,
-    instruction: AssociatedTokenAccountInstruction,
-) -> Instruction {
-    let associated_account_address = get_associated_token_address_with_program_id(
-        wallet_address,
-        token_mint_address,
-        token_program_id,
-    );
-    // safety check, assert if not a creation instruction
-    assert_matches!(
-        instruction,
-        AssociatedTokenAccountInstruction::Create
-            | AssociatedTokenAccountInstruction::CreateIdempotent
-    );
-    Instruction {
-        program_id: id(),
-        accounts: vec![
-            AccountMeta::new(*funding_address, true),
-            AccountMeta::new(associated_account_address, false),
-            AccountMeta::new_readonly(*wallet_address, false),
-            AccountMeta::new_readonly(*token_mint_address, false),
-            AccountMeta::new_readonly(solana_program::system_program::id(), false),
-            AccountMeta::new_readonly(*token_program_id, false),
-        ],
-        data: borsh::to_vec(&instruction).unwrap(),
-    }
-}
-
-/// Creates Create instruction
-pub fn create_associated_token_account(
-    funding_address: &Pubkey,
-    wallet_address: &Pubkey,
-    token_mint_address: &Pubkey,
-    token_program_id: &Pubkey,
-) -> Instruction {
-    build_associated_token_account_instruction(
-        funding_address,
-        wallet_address,
-        token_mint_address,
-        token_program_id,
-        AssociatedTokenAccountInstruction::Create,
-    )
-}
-
-/// Creates CreateIdempotent instruction
-pub fn create_associated_token_account_idempotent(
-    funding_address: &Pubkey,
-    wallet_address: &Pubkey,
-    token_mint_address: &Pubkey,
-    token_program_id: &Pubkey,
-) -> Instruction {
-    build_associated_token_account_instruction(
-        funding_address,
-        wallet_address,
-        token_mint_address,
-        token_program_id,
-        AssociatedTokenAccountInstruction::CreateIdempotent,
-    )
-}
-
-/// Creates a `RecoverNested` instruction
-pub fn recover_nested(
-    wallet_address: &Pubkey,
-    owner_token_mint_address: &Pubkey,
-    nested_token_mint_address: &Pubkey,
-    token_program_id: &Pubkey,
-) -> Instruction {
-    let owner_associated_account_address = get_associated_token_address_with_program_id(
-        wallet_address,
-        owner_token_mint_address,
-        token_program_id,
-    );
-    let destination_associated_account_address = get_associated_token_address_with_program_id(
-        wallet_address,
-        nested_token_mint_address,
-        token_program_id,
-    );
-    let nested_associated_account_address = get_associated_token_address_with_program_id(
-        &owner_associated_account_address, // ATA is wrongly used as a wallet_address
-        nested_token_mint_address,
-        token_program_id,
-    );
-
-    let instruction_data = AssociatedTokenAccountInstruction::RecoverNested;
-
-    Instruction {
-        program_id: id(),
-        accounts: vec![
-            AccountMeta::new(nested_associated_account_address, false),
-            AccountMeta::new_readonly(*nested_token_mint_address, false),
-            AccountMeta::new(destination_associated_account_address, false),
-            AccountMeta::new_readonly(owner_associated_account_address, false),
-            AccountMeta::new_readonly(*owner_token_mint_address, false),
-            AccountMeta::new(*wallet_address, true),
-            AccountMeta::new_readonly(*token_program_id, false),
-        ],
-        data: borsh::to_vec(&instruction_data).unwrap(),
-    }
 }
